Move `BitMask` into `Operators` lint pass
